% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcmm.R
\name{vcmm}
\alias{vcmm}
\title{Model-Based Clustering with Vine Copulas}
\usage{
vcmm(
  data,
  total_comp,
  is_cvine = NA,
  vinestr = NA,
  trunclevel = 1,
  mar = NA,
  bicop = NA,
  methods = c("kmeans"),
  threshold = 1e-04,
  maxit = 10,
  cores = 1
)
}
\arguments{
\item{data}{A matrix or data frame of observations. Categorical/discrete variables not (yet) allowed.
If a matrix or data frame, rows correspond to observations (i) and columns correspond to variables (p).}

\item{total_comp}{An integer specifying the numbers of mixture components (clusters)}

\item{is_cvine}{An integer specifying if the type of components' vine tree structure is C-vine
before/after the ECM phase of clustering.
\itemize{
\item 0 = R-vine (default)
\item 1 = C-vine
}}

\item{vinestr}{A matrix specifying vine tree structures before/after the ECM phase of clustering.
The default is automatic selection.
\link[rvinecopulib]{rvine_structure} checks for a valid R-vine matrix.}

\item{trunclevel}{An integer showing the level of truncation for vine tree structures before the ECM phase of clustering.
The default is 1.}

\item{mar}{A vector of character strings indicating the parametric univariate marginal distributions
to be fitted before/after the ECM phase of clustering.
The default is c('cauchy','gamma','llogis','lnorm','logis','norm','snorm','std', 'sstd').
Other distributions not (yet) allowed.}

\item{bicop}{A vector of strings or integers denoting the bivariate copula families to be fitted
before/after the ECM phase of clustering.
You may use string assignments such as c("all"), c("parametric"), c("nonparametric"), or specific families like c("clayton", "gumbel"). Legacy integer vectors are also accepted and mapped correctly.
The default is c(1,2,3,4,5,6,7,8,10,13,14,16,17,18,20,23,24,26,27,28,30,33,34,36,37,38,40).
\link[rvinecopulib]{bicop_dist} describes the available families with their specifications.}

\item{methods}{A vector of character strings indicating initial clustering method(s) to have a partition
for model selection before the ECM phase of clustering. Current options:
\itemize{
\item 'kmeans' (default)
\item c('kmeans', 'gmm', 'hcVVV')
}}

\item{threshold}{A numeric, stopping the ECM phase of clustering. The default is 1e-4.}

\item{maxit}{An integer, specifying the maximum number of iterations in the CM-step 2 optimization. The default is 10.}

\item{cores}{An integer, showing the number of cores to use for parallel computing.}
}
\value{
An object of class vcmm result. It contains the elements
\describe{
\item{cluster}{the vector with the classification of observations}
\item{output}{a list containing the fitted VCMM.}
} Use \code{print.vcmm_res()} to obtain log-likelihood, BIC, ICL, number of estimated parameters, initial clustering method used
and total number of ECM iterations for the fitted VCMM. \code{summary.vcmm_res()} shows the fitted vine tree structures and
univariate marginal distributions, bivariate copula families with the estimated parameters, as well as
mixture proportions of each component.
}
\description{
It fits vine copula based mixture model distributions to the continuous data
for a given number of components as described in Sahin and Czado (2021)
and use its results for clustering.
}
\examples{
\dontrun{
# Example: fit parametric 4 dimensional vine copula based mixture model with 2 components
# data from UCI Machine Learning Repository 
data_wisc <- read.csv("http://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/wdbc.data", header = FALSE)

# Fit the model
fit <- vcmm(data=data_wisc[,c(15,27,29,30)], total_comp=2, verbose=FALSE)

# Display model statistics
print(fit)

# Extract vine tree structure information
summary(fit)

# Evaluate the density of the fitted model at a given point
RVMs_fitted <- fit$output$vine_models
dens <- dvcmm(c(2.747, 0.1467, 0.13, 0.05334), fit$output$margin, 
              fit$output$marginal_param, RVMs_fitted, fit$output$mixture_prob)
}
}
\references{
Sahin and Czado (2021), Vine copula mixture models and clustering for non-Gaussian data, Econometrics and Statistics.
doi: 10.1016/j.ecosta.2021.08.011
}
\seealso{
\code{\link[=dvcmm]{dvcmm()}}, \code{\link[=rvcmm]{rvcmm()}}
}
